pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "telethon",
    "python-dotenv",
]

[dependency-groups]
dev = [
    "ruff>=0.12.0",
]
test = [
    "pytest>=8.4.1",
    "httpx>=0.28.1",
    "aiogram>=3.20.0.post0",
    "pytest-rerunfailures>=13.0", # Added pytest-rerunfailures
]

[tool.uv.workspace]
members = ["clients/python-client"]
